Added `GenCode`
